# Improve performance of File#each_line iterator.