@asterite 2015-06-08
This is an excellent opportunity to discuss File :-)

I think in Ruby File is always buffered. We can see this with this program:

```crystal
time = Time.now
10_000.times do
  file = File.open(__FILE__)
  while line = file.gets
  end
end
puts Time.now - time
```

It's 0.75s in Crystal vs 0.14s in Ruby.

I'm not sure but maybe all Ruby's IO are internally buffered (Socket, etc.).

In Java (and probably other languages) you are the one to choose if one want buffered IO or not, and this is the approach we took in Crystal. But I'm not sure this is a good choice, because in most cases you do want to buffer things.

So, what should we do?

/cc @waj @bcardiff