@jhass 2015-06-08
+1 to Buffer by default, if you want to offer choice add `UnbufferedIO`