@asterite 2015-06-08
Maybe we can make `File` not be a `FileDescriptorIO` but instead have one and a `BufferedIO` to wrap it and forward methods to it. We could do the same with `Socket` if we wanted to.