@PragTob 2015-06-16
:+1: to buffer by default - sane defaults are great. Plus I still have nightmares from Java IO and wrapping stuff in BufferedReader