@asterite 2015-06-18
@datanoise I just made a commit that makes `FileDescriptorIO` buffered. I think this will simplify things **a lot**. There's also `sync=`/`sync?`.

The original example:

```crystal
puts File.open("/usr/share/dict/words").each_line.max_by{|w| w.length}
```

Before: 1.351s
Now: 0.029s

:smile_cat: